Add a db seed task
