+++ bin/db_seed
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+/app/bin/video_tutorials eval "Elixir.VideoTutorials.Tasks.DbSeed.exec()"
